lxd/network/network/utils: Moves bridge related functions into own file
